Refactor route map
