+++ console/src/modules/system/plugins/PluginDetail.vue
@@ -1,204 +1,148 @@
 <script lang="ts" setup>
-import {
-  VAlert,
-  VDescription,
-  VDescriptionItem,
-  VSwitch,
-} from "@halo-dev/components";
-import type { Ref } from "vue";
-import { computed, inject } from "vue";
+// core libs
+import { provide, ref, computed, onMounted } from "vue";
+import { useRoute } from "vue-router";
 import { apiClient } from "@/utils/api-client";
-import type { Plugin, Role } from "@halo-dev/api-client";
-import { pluginLabels, roleLabels } from "@/constants/labels";
-import { rbacAnnotations } from "@/constants/annotations";
-import { usePluginLifeCycle } from "./composables/use-plugin";
-import { formatDatetime } from "@/utils/date";
+
+// libs
+import cloneDeep from "lodash.clonedeep";
+
+// components
+import { VCard, VPageHeader, VTabbar, VAvatar } from "@halo-dev/components";
+
+// types
+import type { Ref } from "vue";
+import type { Plugin, Setting, SettingForm } from "@halo-dev/api-client";
+import { usePermission } from "@/utils/permission";
+import { useI18n } from "vue-i18n";
 import { useQuery } from "@tanstack/vue-query";
+import type { PluginTab } from "@halo-dev/console-shared";
+import { usePluginModuleStore } from "@/stores/plugin";
+import { markRaw } from "vue";
+import DetailTab from "./tabs/Detail.vue";
+import SettingTab from "./tabs/Setting.vue";
+import { useRouteQuery } from "@vueuse/router";
 
-const plugin = inject<Ref<Plugin | undefined>>("plugin");
-const { changeStatus } = usePluginLifeCycle(plugin);
+const { currentUserHasPermission } = usePermission();
+const { t } = useI18n();
 
-interface RoleTemplateGroup {
-  module: string | null | undefined;
-  roles: Role[];
-}
-
-const { data: pluginRoleTemplates } = useQuery({
-  queryKey: ["plugin-roles", plugin?.value?.metadata.name],
-  queryFn: async () => {
-    const { data } = await apiClient.extension.role.listv1alpha1Role({
-      page: 0,
-      size: 0,
-      labelSelector: [
-        `${pluginLabels.NAME}=${plugin?.value?.metadata.name}`,
-        `${roleLabels.TEMPLATE}=true`,
-        "!halo.run/hidden",
-      ],
-    });
-
-    return data.items;
+const initialTabs = ref<PluginTab[]>([
+  {
+    id: "detail",
+    label: t("core.plugin.tabs.detail"),
+    component: markRaw(DetailTab),
   },
-  cacheTime: 0,
-  enabled: computed(() => !!plugin?.value?.metadata.name),
+]);
+
+const route = useRoute();
+
+const tabs = ref<PluginTab[]>(cloneDeep(initialTabs.value));
+const activeTab = useRouteQuery<string>("tab", tabs.value[0].id, {
+  mode: "push",
 });
 
-const pluginRoleTemplateGroups = computed<RoleTemplateGroup[]>(() => {
-  const groups: RoleTemplateGroup[] = [];
-  pluginRoleTemplates.value?.forEach((role) => {
-    const group = groups.find(
-      (group) =>
-        group.module === role.metadata.annotations?.[rbacAnnotations.MODULE]
-    );
-    if (group) {
-      group.roles.push(role);
-    } else {
-      groups.push({
-        module: role.metadata.annotations?.[rbacAnnotations.MODULE],
-        roles: [role],
+provide<Ref<string | undefined>>("activeTab", activeTab);
+
+const { data: plugin } = useQuery({
+  queryKey: ["plugin", route.params.name],
+  queryFn: async () => {
+    const { data } =
+      await apiClient.extension.plugin.getpluginHaloRunV1alpha1Plugin({
+        name: route.params.name as string,
       });
+    return data;
+  },
+});
+
+provide<Ref<Plugin | undefined>>("plugin", plugin);
+
+const { data: setting } = useQuery({
+  queryKey: ["plugin-setting", plugin],
+  queryFn: async () => {
+    const { data } = await apiClient.plugin.fetchPluginSetting({
+      name: plugin.value?.metadata.name as string,
+    });
+    return data;
+  },
+  enabled: computed(() => {
+    return (
+      !!plugin.value &&
+      !!plugin.value.spec.settingName &&
+      currentUserHasPermission(["system:plugins:manage"])
+    );
+  }),
+  async onSuccess(data) {
+    if (data) {
+      const { forms } = data.spec;
+      tabs.value = [
+        ...initialTabs.value,
+        ...forms.map((item: SettingForm) => {
+          return {
+            id: item.group,
+            label: item.label || "",
+            component: markRaw(SettingTab),
+          };
+        }),
+      ] as PluginTab[];
+    }
+  },
+});
+
+provide<Ref<Setting | undefined>>("setting", setting);
+
+onMounted(() => {
+  const { pluginModules } = usePluginModuleStore();
+  const currentPluginModule = pluginModules.find(
+    (item) => item.extension.metadata.name === route.params.name
+  );
+
+  if (!currentPluginModule) {
+    return;
+  }
+
+  const { extensionPoints } = currentPluginModule;
+
+  if (!extensionPoints?.["plugin:tabs:create"]) {
+    return;
+  }
+
+  const extraTabs = extensionPoints["plugin:tabs:create"]() as PluginTab[];
+
+  extraTabs.forEach((tab) => {
+    if (currentUserHasPermission(tab.permissions)) {
+      initialTabs.value.push(tab);
     }
   });
-  return groups;
 });
 </script>
-
 <template>
-  <Transition mode="out-in" name="fade">
-    <div>
-      <div class="flex items-center justify-between bg-white px-4 py-4 sm:px-6">
-        <div>
-          <h3 class="text-lg font-medium leading-6 text-gray-900">
-            {{ $t("core.plugin.detail.header.title") }}
-          </h3>
-        </div>
-        <div v-permission="['system:plugins:manage']">
-          <VSwitch :model-value="plugin?.spec.enabled" @change="changeStatus" />
-        </div>
+  <VPageHeader :title="plugin?.spec?.displayName">
+    <template #icon>
+      <VAvatar
+        v-if="plugin"
+        :src="plugin.status?.logo"
+        :alt="plugin.spec.displayName"
+        class="mr-2"
+        size="sm"
+      />
+    </template>
+  </VPageHeader>
+
+  <div class="m-0 md:m-4">
+    <VCard :body-class="['!p-0']">
+      <template #header>
+        <VTabbar
+          v-model:active-id="activeTab"
+          :items="tabs.map((item) => ({ id: item.id, label: item.label }))"
+          class="w-full !rounded-none"
+          type="outline"
+        ></VTabbar>
+      </template>
+      <div class="bg-white">
+        <template v-for="tab in tabs" :key="tab.id">
+          <component :is="tab.component" v-if="activeTab === tab.id" />
+        </template>
       </div>
-      <div
-        v-if="
-          plugin?.status?.phase === 'FAILED' &&
-          plugin?.status?.conditions?.length
-        "
-        class="w-full px-4 pb-2 sm:px-6"
-      >
-        <VAlert
-          type="error"
-          :title="plugin?.status?.conditions?.[0].reason"
-          :description="plugin?.status?.conditions?.[0].message"
-          :closable="false"
-        />
-      </div>
-      <div class="border-t border-gray-200">
-        <VDescription>
-          <VDescriptionItem
-            :label="$t('core.plugin.detail.fields.display_name')"
-            :content="plugin?.spec.displayName"
-          />
-          <VDescriptionItem
-            :label="$t('core.plugin.detail.fields.description')"
-            :content="plugin?.spec.description"
-          />
-          <VDescriptionItem
-            :label="$t('core.plugin.detail.fields.version')"
-            :content="plugin?.spec.version"
-          />
-          <VDescriptionItem
-            :label="$t('core.plugin.detail.fields.requires')"
-            :content="plugin?.spec.requires"
-          />
-          <VDescriptionItem :label="$t('core.plugin.detail.fields.author')">
-            <a
-              v-if="plugin?.spec.author"
-              :href="plugin?.spec.author.website"
-              target="_blank"
-            >
-              {{ plugin?.spec.author.name }}
-            </a>
-            <span v-else>
-              {{ $t("core.common.text.none") }}
-            </span>
-          </VDescriptionItem>
-          <VDescriptionItem :label="$t('core.plugin.detail.fields.license')">
-            <ul
-              v-if="plugin?.spec.license && plugin?.spec.license.length"
-              class="list-inside"
-              :class="{ 'list-disc': plugin?.spec.license.length > 1 }"
-            >
-              <li v-for="(license, index) in plugin.spec.license" :key="index">
-                <a v-if="license.url" :href="license.url" target="_blank">
-                  {{ license.name }}
-                </a>
-                <span v-else>
-                  {{ license.name }}
-                </span>
-              </li>
-            </ul>
-          </VDescriptionItem>
-          <VDescriptionItem
-            :label="$t('core.plugin.detail.fields.role_templates')"
-          >
-            <dl
-              v-if="pluginRoleTemplateGroups.length"
-              class="divide-y divide-gray-100"
-            >
-              <div
-                v-for="(group, groupIndex) in pluginRoleTemplateGroups"
-                :key="groupIndex"
-                class="rounded bg-gray-50 px-4 py-5 hover:bg-gray-100 sm:grid sm:grid-cols-3 sm:gap-4 sm:px-6"
-              >
-                <dt class="text-sm font-medium text-gray-900">
-                  {{ group.module }}
-                </dt>
-                <dd class="mt-1 text-sm text-gray-900 sm:col-span-2 sm:mt-0">
-                  <ul class="space-y-2">
-                    <li v-for="(role, index) in group.roles" :key="index">
-                      <div
-                        class="inline-flex w-72 cursor-pointer flex-row items-center gap-4 rounded border p-5 hover:border-primary"
-                      >
-                        <div class="inline-flex flex-col gap-y-3">
-                          <span class="font-medium text-gray-900">
-                            {{
-                              role.metadata.annotations?.[
-                                rbacAnnotations.DISPLAY_NAME
-                              ]
-                            }}
-                          </span>
-                          <span
-                            v-if="
-                              role.metadata.annotations?.[
-                                rbacAnnotations.DEPENDENCIES
-                              ]
-                            "
-                            class="text-xs text-gray-400"
-                          >
-                            {{
-                              $t("core.role.common.text.dependent_on", {
-                                roles: JSON.parse(
-                                  role.metadata.annotations?.[
-                                    rbacAnnotations.DEPENDENCIES
-                                  ]
-                                ).join(", "),
-                              })
-                            }}
-                          </span>
-                        </div>
-                      </div>
-                    </li>
-                  </ul>
-                </dd>
-              </div>
-            </dl>
-            <span v-else>
-              {{ $t("core.common.text.none") }}
-            </span>
-          </VDescriptionItem>
-          <VDescriptionItem
-            :label="$t('core.plugin.detail.fields.last_starttime')"
-            :content="formatDatetime(plugin?.status?.lastStartTime)"
-          />
-        </VDescription>
-      </div>
-    </div>
-  </Transition>
+    </VCard>
+  </div>
 </template>
